Remove task from 'guix-daemon.conf'.
The 'task' means that events that led to this job starting will be blocked
until it has stopped. Tasks are short lived jobs, whereas the guix-daemon is a
long lived service. Including 'task' means that attempts to start the
guix-daemon appear to hang, as upstart waits for it to exit.

* etc/guix-daemon.conf.in: Remove 'task'.