DRAFT system: Add (gnu system bootstrap).
This allows us to perform arbitrary builds on a system that has no
userland besides the build process itself, running as PID 1.

Suggested by Vagrant Cascadian.

DRAFT: The resulting system does build things, but this is all happening
into memory, which may or may not be a problem (it allows us to not have
disk drivers in the kernel!).  More importantly, it does not display
anything upon completion, and the build result is lost as well.

* gnu/system/bootstrap.scm: New file.
* gnu/local.mk (GNU_SYSTEM_MODULES): Add it.