DRAFT installer: Implement a dialog on /var/guix/installer-socket.
DRAFT We might need to instrument some more pages.

This will allow us to automate testing of the installer.

* gnu/installer/utils.scm (%client-socket-file)
(current-server-socket, current-clients): New variables.
(open-server-socket, call-with-server-socket): New procedure.
(with-server-socket): New macro.
(run-shell-command): Add call to 'send-to-clients'.  Select on both
current-input-port and current-clients.
* gnu/installer/steps.scm (run-installer-steps): Wrap 'call-with-prompt'
in 'with-socket-server'.
* gnu/installer/newt/page.scm (watch-clients!, call-with-client): New
procedures.
(with-client): New macro.
(send-to-clients): New procedures.
(draw-info-page): Add call to 'send-to-clients'.
(run-input-page): Add calls to 'watch-clients!' and 'send-to-clients'.
Handle EXIT-REASON equal to 'exit-fd-ready.
(run-confirmation-page): Likewise.
(run-listbox-selection-page): Add #:client-callback-procedure
parameter.  Wrap body in 'loop'.  Add calls to 'watch-clients!' and
'send-to-clients'.  Define 'choice->item' and use it.
(run-checkbox-tree-page): Likewise.
(run-file-textbox-page): Add calls to 'watch-clients!' and
'send-to-clients'.  Handle 'exit-fd-ready'.
* gnu/installer/newt/partition.scm (run-disk-page): Pass
 #:client-callback-procedure to 'run-listbox-selection-page'.
* gnu/installer/newt/user.scm (run-user-page): Add calls to
'watch-clients!' and 'send-to-clients'.  Handle 'exit-fd-ready'.
* gnu/installer/newt/welcome.scm (run-menu-page): Define
'choice->item' and use it.  Add calls to 'watch-clients!' and
'send-to-clients'.  Wrap body in 'loop'.