gnu: texlive-amsfonts/fixed: Do not generate tfm from afm and pl files.
The TFM files are included in the amsfonts bundle and are meant to installed
as is.  They were not meant to be the result of converting afm or pl files.
We ended up with subtly broken fonts due to this attempted conversion.

* gnu/packages/tex.scm (texlive-amsfonts/fixed)[source]: Fetch tfm directory.
[arguments]: Remove custom #:modules; remove build phases 'build-fonts and
'install-generated-fonts; adjust 'remove-extra-files; add build phase 'chdir
and 'unchdir.
[native-inputs]: Remove texlive-fontinst.