gnu: upower: Update to 0.99.15.
Upstream has removed the Autotools build system and now
only supports Meson, so the build system has been changed
to 'meson-build-system'.

The patch 'power-builddir.patch' has to be replaced to
adjust to the switch to Meson.  This has been done by adding
an origin snippet.

The origin snippet has been simplified, and for consistency,
it has been made a gexp.

Some dependencies were Autotools-only and could be removed.

* gnu/packages/gnome.scm (upower): Update to 0.99.15.

Signed-off-by: Ludovic Courtès <ludo@gnu.org>