gnu: java-xpp3: Download source from ci.guix.gnu.org.
The www.extreme.indiana.edu site is now serving a redirect:

  This site has been deprecated. You can access an archive of this site at
  https://web.archive.org/web/20210225153105/https://www.extreme.indiana.edu/.

Because this isn't a 404, as it probably should be, guix follows this redirect
and gets some content that isn't the file it asked for.

To avoid this incorrect behaviour, switch to using a ci.guix.gnu.org URL for
the file instead.

* gnu/packages/xml.scm (java-xpp3)[source]: Switch the uri to ci.guix.gnu.org.