gnu: cl-slime-swank: Store .fasl in the ASDF registry.
* gnu/packages/lisp-xyz.scm (sbcl-slime-swank)[arguments]: Patch swank.asd to
store the .fasl in the ASDF registry so that they are found in the
package.
Also patch swank-loader.lisp so that it does not try to remove the .fasl from
the store.

(cl-slime-swank)[arguments]: Revert above patch of the swank.ask so that the
source package does not get bound to the SBCL package.

We would not have to do this if SLIME/Swank used ASDF properly instead of
working around it.