xdg-utils: Support cross-compilation.
It has been verified that this does not cause rebuilds when compiling
natively.  The references graph when cross-compiling has also been verified --
glibc-2.33 and the native bash-static-5.1.8 still remains in the graph, but
via the cross-compiled inetutils-2.0, ncurses-6.2.20210619 and via
gcc-cross-TARGET-10.3.0-lib, which is not related with Perl cross-compilation.

* gnu/packages/freedesktop.scm
(xdg-utils)[inputs]{bash-minimal,file}: New inputs when cross-compiling.
(xdg-utils)[arguments]<#:phases>{locate-catalog-files}: Add 'native-inputs'
argument when cross-compiling. Look for docbook-xml and docbook-xsl in
native-inputs when cross-compiling.  While we are at it, eliminate input
labels with search-input-file.
(xdg-utils)[arguments]<#:phases>{patch-hardcoded-patch}: Use
search-input-file + inputs instead of 'which' when cross-compiling.

Signed-off-by: Mathieu Othacehe <othacehe@gnu.org>