gnu: samba: Update to 4.16.4.
As a reproducibility check, you can check the hash:

$ guix hash -r $(./pre-inst-env guix build samba --no-grafts --system=x86_64-linux)
1mg3nh705pbiz3ihq8jzkmjb15m6y8zjnzyr4hxmbciv7frgykk0

I've checked that it's correctly signed, and I've tried to verify that the PGP
key is correct -- searching for 'Samba Distribution Verification Key', the
websites of the first search results have the same key:

* https://www.linuxcompatible.org/story/samba-4150rc5-released/
* https://wiki.samba.org/index.php/Samba_4.12_Features_added/changed
* https://www.mail-archive.com/samba-announce@lists.samba.org/msg00562.html
* https://wiki.linuxfromscratch.org/blfs/ticket/14748

https://groups.google.com/g/linux.samba/c/k3xaN0TIcQU has a different key,
but as mentioned on the previous locations, the key has changed.

I have also checked the source code diff (using diffoscope) between the the
current version and the previous, nothing appears 'suspicious' (with the
caveat that it is written in C).

According to "guix refresh -l" there aren't too many dependents.  I'm
currently rebuilding them, no failures so far except for lxde, but it already
failed to build before this patch.

* gnu/packages/samba.scm (samba): Update to 4.16.4. To help future updaters,
add the current PGP fingerprint in a comment.

Signed-off-by: Marius Bakke <marius@gnu.org>