gnu: Add comments to the bottom of a few more package modules.
This will hopefully discourage patches which add new packages to the bottom of
files, as this increases the likelihood of conflicts when applying patches.

If there are more specific ways that packages should be arranged in specific
modules, this message can be updated accordingly.

* gnu/packages/crates-graphics.scm: Add comment discouraging adding packages
to the bottom of the file.
* gnu/packages/crates-gtk.scm: ditto.
* gnu/packages/games.scm: ditto.
* gnu/packages/music.scm: ditto.
* gnu/packages/qt.scm: ditto.
* gnu/packages/web.scm: ditto.