build: jami-service: account->username always return a fingerprint.
This change is motivated by the fact that Account.registeredName is a volatile
account data, not exported along the account and retrieved from the name
server.  Have it always return Account.username instead, so that the result is
reproducible independent of whether networking is available or not.

* gnu/build/jami-service.scm (account->username): Always return the account
fingerprint.  Adjust doc.
(id->username): Likewise.