gnu: Add go-webring.
* gnu/packages/web.scm (go-webring): New variable.

Hi paren,

Thanks for the review. I changed everything but the blog stuff because
go-webring is not necessarily used in the context of a blog. A webring
can stand on it's own without being associated with a blog. A webring
can associate websites that are not blogs or Guix channels ;()

Here's v2

all best,

jgart

Signed-off-by: Ludovic Courtès <ludo@gnu.org>