tests: xvnc: Do not use specification->package in OS definition.
Doing so would cause the extra package dependencies to not be
correctly registered as dependencies, which would lead to a silent
failure when attempting to load or byte compile the module at the time
guix build itself (e.g., when running 'guix pull').

* gnu/tests/vnc.scm (%xvnc-os) [packages]: Turn the the
specification->package declaration into a list of package objects.