gnu: ghc-cryptohash-sha256: Update cabal revision metadata.
Before this commit, the .cabal metadata file referred to the previous version,
but was not noticed because it had no incompatible changes, and was 404
upstream (thus falling back to cache) until recently a new metadata revision
was made for 0.11.102.1.

  https://hackage.haskell.org/package/cryptohash-sha256-0.11.102.1/revisions/

This is a follow-up to commit 54f4502e459cf39512e30cbed4fac17c0aabe79d, which
should have reset #:cabal-revision to 0 and changed the hash.

* gnu/packages/haskell-crypto.scm (ghc-cryptohash-sha256)[arguments]: Update
 #:cabal-revision hash.