gnu: po4a: Update to 0.68
The phase do-not-override-PERL5LIB is removed since it was fixed in po4a v0.63.
https://github.com/mquinson/po4a/blob/0ab1670e50f0a72781e3d1de6ab9da0c2d71c646/NEWS#L366

bash-minimal is added to inputs since the wrap-program function is used.
See also https://issues.guix.gnu.org/49327#80

The only regrettable thing is that only the SGML test failed, but I could not
find the cause...  Nevertheless, some previously failed tests are now passing.

* gnu/packages/gettext.scm (po4a): Update to 0.68.

Signed-off-by: Julien Lepiller <julien@lepiller.eu>