gnu: gnulib: Refer to "gcc" as provided via implicit package inputs.
Fixes <https://issues.guix.gnu.org/60947>.

Previously, with the #+(file-append gcc "/bin/gcc") reference, the
resulting derivation would end up referring to the grafted GCC when
grafts are enabled, and to the ungrafted one otherwise.  As a result, a
different derivation would be produced depending on whether grafts are
enabled.

* gnu/packages/build-tools.scm (gnulib-checkout)[arguments]: Refer to
"gcc", not #+(file-append gcc "/bin/gcc").